+++ besu/src/test/java/org/hyperledger/besu/cli/options/stable/DataStorageOptionsTest.java
@@ -34,7 +34,7 @@ public class DataStorageOptionsTest
         dataStorageConfiguration ->
             assertThat(dataStorageConfiguration.getUnstable().getBonsaiTrieLogPruningWindowSize())
                 .isEqualTo(600),
-        "--Xbonsai-limit-trie-logs-enabled",
+        "--bonsai-limit-trie-logs-enabled",
         "--Xbonsai-trie-logs-pruning-window-size",
         "600");
   }
@@ -45,14 +45,14 @@ public class DataStorageOptionsTest
         dataStorageConfiguration ->
             assertThat(dataStorageConfiguration.getUnstable().getBonsaiLimitTrieLogsEnabled())
                 .isEqualTo(false),
-        "--Xbonsai-limit-trie-logs-enabled=false");
+        "--bonsai-limit-trie-logs-enabled=false");
   }
 
   @Test
   public void bonsaiTrieLogPruningWindowSizeShouldBePositive() {
     internalTestFailure(
         "--Xbonsai-trie-logs-pruning-window-size=0 must be greater than 0",
-        "--Xbonsai-limit-trie-logs-enabled",
+        "--bonsai-limit-trie-logs-enabled",
         "--Xbonsai-trie-logs-pruning-window-size",
         "0");
   }
@@ -61,7 +61,7 @@ public class DataStorageOptionsTest
   public void bonsaiTrieLogPruningWindowSizeShouldBeAboveRetentionLimit() {
     internalTestFailure(
         "--Xbonsai-trie-logs-pruning-window-size=127 must be greater than --bonsai-historical-block-limit=512",
-        "--Xbonsai-limit-trie-logs-enabled",
+        "--bonsai-limit-trie-logs-enabled",
         "--Xbonsai-trie-logs-pruning-window-size",
         "127");
   }
@@ -72,7 +72,7 @@ public class DataStorageOptionsTest
         dataStorageConfiguration ->
             assertThat(dataStorageConfiguration.getBonsaiMaxLayersToLoad())
                 .isEqualTo(MINIMUM_BONSAI_TRIE_LOG_RETENTION_LIMIT + 1),
-        "--Xbonsai-limit-trie-logs-enabled",
+        "--bonsai-limit-trie-logs-enabled",
         "--bonsai-historical-block-limit",
         "129");
   }
@@ -83,7 +83,7 @@ public class DataStorageOptionsTest
         dataStorageConfiguration ->
             assertThat(dataStorageConfiguration.getBonsaiMaxLayersToLoad())
                 .isEqualTo(MINIMUM_BONSAI_TRIE_LOG_RETENTION_LIMIT),
-        "--Xbonsai-limit-trie-logs-enabled",
+        "--bonsai-limit-trie-logs-enabled",
         "--bonsai-historical-block-limit",
         "128");
   }
@@ -92,7 +92,7 @@ public class DataStorageOptionsTest
   public void bonsaiTrieLogRetentionLimitShouldBeAboveMinimum() {
     internalTestFailure(
         "--bonsai-historical-block-limit minimum value is 128",
-        "--Xbonsai-limit-trie-logs-enabled",
+        "--bonsai-limit-trie-logs-enabled",
         "--bonsai-historical-block-limit",
         "127");
   }